requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.1
typing_extensions==4.15.0
uvicorn==0.35.0
Werkzeug==3.1.3
